safari-534.53-branch/LayoutTests/fast/tokenizer/unterminated-quote-in-noscript.html

<html>
<body>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();
</script>
<noscript> 
<img src="about:blank" alt=">
</noscript>
<div>PASS: This text is visible.</div>
</body>
</html>
